Fixed report typos
This is also a prelude to the creation of a timer for the questions. I
moved the questionText code out of the Game function because when I
tried creating a timer earlier, the questionText was constantly getting
re-rendered and re-created every tick of a second. This was causing the
questionText to randomly change sometimes due to one of the
introVariations getting chosen.

Hopefully my code here will prevent that problem with future timer
functionality.